% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task6Lab5.R
\name{check_p_binom}
\alias{check_p_binom}
\title{Function to grab the p binomial}
\usage{
check_p_binom(success, trials, prob)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
p_binom(success, trials, prob)
}
\description{
Function to grab the p binomial
}
