% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{World Health Organization TB data}
\format{
## `who`
A data frame with 15 rows and 2 variables:
\describe{
\item{DISTANCE}{Distanace from a fire station in miles}
\item{DAMAGE}{X 1000 dollars in damgae}
...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
fire
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
