% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Function to create and plot confidence intervals for a number of samples using different functions}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{the number of times to repeat the number of samples}

\item{x}{the samples}

\item{fun}{the function you want to calculate}

\item{alpha}{used to generate the desired confidence interval}
}
\value{
a list containing the confidence interval ci, xstat (iter samples), x (the samples), the function
}
\description{
Function to create and plot confidence intervals for a number of samples using different functions
}
