% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Performs a t.test to calculate the confidence interval for the mean from a random sample}
\usage{
myci(x)
}
\arguments{
\item{x}{is the value of samples}
}
\description{
Performs a t.test to calculate the confidence interval for the mean from a random sample
}
