% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cltPoisson.R
\name{mycltp}
\alias{mycltp}
\title{This function shows the application of the central limit theorem
on the Poisson distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{total number of samples}

\item{iter}{the number of times to iterate per sample}

\item{lambda}{the average number of events in the time interval}
}
\description{
This function shows the application of the central limit theorem
on the Poisson distribution
}
