% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculates the number of tickets sold  when there are N seats available,
p is the probability a person will show, and gamma is the probability
a plane will be truly overbooked}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{total number of seats available}

\item{gamma}{the probability that a plane will be overbooked}

\item{p}{is the probability that a person will show}
}
\description{
Calculates the number of tickets sold  when there are N seats available,
p is the probability a person will show, and gamma is the probability
a plane will be truly overbooked
}
