% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{Performs a t.test to calculate the confidence interval for the mean of two random samples.
Calculates the tcalc, pvalue, and plots the information}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{x1}{is the first sample}

\item{x2}{is the second sample}
}
\description{
Performs a t.test to calculate the confidence interval for the mean of two random samples.
Calculates the tcalc, pvalue, and plots the information
}
