% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_package.R
\docType{package}
\name{functions_package}
\alias{Math4753Hari}
\alias{Math4753Hari-package}
\alias{functions_package}
\title{Functions package}
\description{
Package for all functions I will use
}
\author{
Harivansh Luchmun
}
