% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Calculates the max likelihood given a parameter and a set of values}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{is the function used to calculate the likelihood}

\item{x}{iss the vector of values}

\item{param}{is the target parameter to find the likelihood}

\item{...}{arguments to change specific parts}
}
\description{
Calculates the max likelihood given a parameter and a set of values
}
