% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Function to plot a normal curve and region while giving the area of the region}
\usage{
myncurve(x, mu, sigma, a = 1)
}
\arguments{
\item{x}{is the vector of values}

\item{mu}{the mean of the distribution}

\item{sigma}{the standard deviation of the distribution}

\item{a}{the value of x you want to evaluate at}
}
\description{
Function to plot a normal curve and region while giving the area of the region
}
